Authorization feature added

+++ src/main/java/com/mpt/journal/controller/EmailConfirmationController.java
@@ -6,11 +6,15 @@ import com.mpt.journal.domain.entity.EmailConfirmationEntity;
 import com.mpt.journal.domain.model.PagedResult;
 import com.mpt.journal.domain.service.EmailConfirmationsService;
 import com.mpt.journal.domain.service.UsersService;
+import org.apache.tomcat.util.net.openssl.ciphers.Authentication;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.sql.Date;
 import java.util.*;
@@ -23,6 +27,8 @@ public class EmailConfirmationController {
     @Autowired
     private UsersService _users;
 
+
+    @PreAuthorize("hasAuthority('USER')")
     @GetMapping("/emailConf")
     public String getEmailConfs(
             Model model,
@@ -54,25 +60,43 @@ public class EmailConfirmationController {
         return "emailConfirmations/emailConfs";
     }
 
+    @PreAuthorize("hasAuthority('USER')")
     @GetMapping("/emailConf/add")
     public String addEmailConf(
             Model model
     ) {
-        var allowedUsers = _users.getUsers(1, Integer.MAX_VALUE, null, null, null, null).getValue();
+        var auth = SecurityContextHolder.getContext().getAuthentication();
+        var authorities = auth.getAuthorities();
+
+
+        var allowedUsers = authorities.stream().anyMatch(role -> role.getAuthority().equals("ADMIN")) ?
+                _users.getUsers(1, Integer.MAX_VALUE, null, null, null, null).getValue()
+                : Collections.singletonList(_users.getUserByLogin(auth.getName()));
+
         model.addAttribute("allowedUsers", allowedUsers);
         model.addAttribute("emailConf_model", new EmailConfirmationEntity());
 
         return "emailConfirmations/createEmailConf";
     }
 
+    @PreAuthorize("hasAuthority('USER')")
     @PostMapping("/emailConf/add")
     public String postEmailConf(
             Model model,
             @ModelAttribute EmailConfirmationEntity emailConf,
-            BindingResult bindingResult
+            BindingResult bindingResult,
+            RedirectAttributes attributes
     ) {
-        if (!bindingResult.hasErrors())
+        if (!bindingResult.hasErrors()) {
+            var auth = SecurityContextHolder.getContext().getAuthentication();
+            if (auth.getAuthorities().stream().noneMatch(role -> role.getAuthority().equals("ADMIN")) &&
+                    emailConf.getUser() != null && !emailConf.getUser().getLogin().equals(auth.getName())) {
+                attributes.addAttribute("error_message", "Not permitted");
+                return "redirect:/emailConf";
+            }
+
             _emailConfirmations.addEmailConfirmation(emailConf);
+        }
 
         return "redirect:/emailConf";
     }
@@ -80,10 +104,23 @@ public class EmailConfirmationController {
     @GetMapping("/emailConf/update/{id}")
     public String editEmailConf(
             Model model,
-            @PathVariable UUID id
+            @PathVariable UUID id,
+            RedirectAttributes attributes
     ) {
         var confirmation = _emailConfirmations.getEmailConfirmationByID(id);
-        var allowedUsers = _users.getUsers(1, Integer.MAX_VALUE, null, null, null, null).getValue();
+        var auth = SecurityContextHolder.getContext().getAuthentication();
+        var authorities = auth.getAuthorities();
+
+        if (confirmation.getUser() != null && !confirmation.getUser().getLogin().equals(auth.getName())) {
+            attributes.addAttribute("error_message", "Not permitted");
+            return "redirect:/emailConf";
+        }
+
+
+        var allowedUsers = authorities.stream().anyMatch(role -> role.getAuthority().equals("ADMIN")) ?
+                _users.getUsers(1, Integer.MAX_VALUE, null, null, null, null).getValue()
+                : Collections.singletonList(_users.getUserByLogin(auth.getName()));
+
         model.addAttribute("allowedUsers", allowedUsers);
         model.addAttribute("emailConf_model", confirmation);
 
@@ -105,8 +142,19 @@ public class EmailConfirmationController {
     @GetMapping("/emailConf/delete/{id}")
     public String deleteEmailConf(
             Model model,
-            @PathVariable UUID id
+            @PathVariable UUID id,
+            RedirectAttributes attributes
     ) {
+        var auth = SecurityContextHolder.getContext().getAuthentication();
+        var confirmation = _emailConfirmations.getEmailConfirmationByID(id);
+
+        if (confirmation != null &&
+                auth.getAuthorities().stream().noneMatch(role -> role.getAuthority().equals("ADMIN")) &&
+                (confirmation.getUser() == null || !confirmation.getUser().getLogin().equals(auth.getName()))){
+            attributes.addAttribute("error_message", "Not permitted");
+            return "redirect:/emailConf";
+        }
+
         _emailConfirmations.deleteEmailConfirmation(id);
         return "redirect:/emailConf";
     }
